+++ hubsmgr/utility/git.py
