+++ hubsmgr/utility/parserutils.py
