+++ hubsmgr/utility/sha1utils.py
